refactor(watch): split W-cycle resolver into WatchModeController.Cycle.cs
